clean up violin by tuning width range, delete "strokeWidth" and expand integration test with ssgsea & samplelst
